+++ ivy.xml
@@ -293,8 +293,13 @@ under the License.
     </dependency>
 
     <dependency org="org.apache.phoenix" name="phoenix-core" rev="${phoenix.version}" conf="phoenix->default">
-      <artifact name="phoenix-core" type="jar"/>
-      <artifact name="phoenix-core" type="test-jar" ext="jar" m:classifier="tests"/>
+       <artifact name="phoenix-core" type="jar"/>
+       <artifact name="phoenix-core" type="test-jar" ext="jar" m:classifier="tests"/>
+       <exclude org="org.apache.hbase" module="hbase-common"/>
+       <exclude org="org.apache.hbase" module="hbase-prefix-tree"/>
+       <exclude org="org.apache.hbase" module="hbase-client"/>
+       <exclude org="org.apache.hbase" module="hbase-server"/>
+       <exclude org="org.apache.hbase" module="hbase-protocol"/>
     </dependency>
     
     <dependency org="org.apache.hive.hcatalog" name="hive-hcatalog-core"
